+++ src/Semantic/Api/Bridge.hs
@@ -15,7 +15,6 @@ import           Analysis.Name
 import           Control.Lens
 import           Data.Bifunctor
 import qualified Data.Blob as Data
-import qualified Data.Edit as Data
 import           Data.Either
 import           Data.Foldable
 import           Data.Functor.Tagged
@@ -96,7 +95,7 @@ instance APIBridge API.Blob Data.Blob where
     blobToApiBlob b
       = defMessage
       & P.content .~ Source.toText (Data.blobSource b)
-      & P.path .~ T.pack (Data.blobPath b)
+      & P.path .~ T.pack (Data.blobFilePath b)
       & P.language .~ (bridging # Data.blobLanguage b)
     apiBlobToBlob blob =
       let src = blob^.P.content.to Source.fromText
@@ -105,71 +104,3 @@ instance APIBridge API.Blob Data.Blob where
       { blobSource = src
       , blobFile = File pth (Source.totalSpan src) (blob^.P.language.bridging)
       }
-
-instance APIConvert API.BlobPair Data.BlobPair where
-  converting = prism' blobPairToApiBlobPair apiBlobPairToBlobPair where
-
-    apiBlobPairToBlobPair blobPair = case (blobPair^.P.maybe'before, blobPair^.P.maybe'after) of
-      (Just before, Just after) -> Just $ Data.Compare (before^.bridging) (after^.bridging)
-      (Just before, Nothing)    -> Just $ Data.Delete (before^.bridging)
-      (Nothing, Just after)     -> Just $ Data.Insert (after^.bridging)
-      _                         -> Nothing
-
-    blobPairToApiBlobPair (Data.Compare before after) = defMessage & P.maybe'before .~ (bridging #? before) & P.maybe'after .~ (bridging #? after)
-    blobPairToApiBlobPair (Data.Insert after)         = defMessage & P.maybe'before .~ Nothing & P.maybe'after .~ (bridging #? after)
-    blobPairToApiBlobPair (Data.Delete before)        = defMessage & P.maybe'before .~ (bridging #? before) & P.maybe'after .~ Nothing
-
--- * Stack graph stuff. This doesn't fit into the optics formulation because it relies on
--- contextual information
-
-data PathConvertError
-  = NodeNotFound Tag
-  | InvalidStartingSize Int64
-  | InvariantViolated SG.PathInvariantError
-
-type NodeCache = IM.IntMap (Tagged SG.Node)
-
-stackDelimiter :: T.Text
-stackDelimiter = T.pack "\x1E"
-
-pathToApiPath :: SG.Path -> API.StackGraphPath
-pathToApiPath path =
-  defMessage
-    & P.from .~ path^.SG.startingNode_.identifier
-    & P.to .~ path^.SG.endingNode_.identifier
-    & P.endingScopeStack .~ path^.SG.endingScopeStack_
-    & P.startingSymbolStack .~ fmap formatName (toList $ path^.SG.startingSymbolStack_)
-    & P.endingSymbolStack .~ fmap formatName (toList $ path^.SG.endingSymbolStack_)
-    & P.startingScopeStackSize .~ (if path^.SG.startingScopeStackSize_ == SG.One then 1 else 0)
-    & P.edges .~ formatEdges (path^.SG.edges_)
-      where
-        -- TODO: It's not clear from my reading of the docs how exactly we want to format
-        -- edge strings.
-        formatEdges = T.intercalate stackDelimiter . toList . fmap SG.formatEdge
-
-apiPathToPath :: NodeCache -> API.StackGraphPath -> Either (NonEmpty PathConvertError) SG.Path
-apiPathToPath nc path = Validation.toEither validate >>= ensureInvariantsHold
-  where
-    validate = do
-      let failing = Validation.Failure . pure @NonEmpty @PathConvertError
-          lookupNode n = case IM.lookup (fromIntegral n) nc of
-            Just a  -> pure a
-            Nothing -> failing (NodeNotFound n)
-          ensureHolds p = Validation.Success p
-
-      startingNode <- lookupNode (path^.P.from)
-      endingNode <- lookupNode (path^.P.to)
-      endingScopeStack <- pure (path^.P.endingScopeStack)
-      edges <- pure . Seq.fromList . SG.parseEdges . view P.edges $ path
-      startingSymbolStack <- pure . Seq.fromList . fmap name . toList . view P.startingSymbolStack $ path
-      endingSymbolStack <- pure . Seq.fromList . fmap name . toList . view P.startingSymbolStack $ path
-      startingScopeStackSize <- case path^.P.startingScopeStackSize of
-        0 -> pure SG.Zero
-        1 -> pure SG.One
-        n -> failing (InvalidStartingSize n)
-      pure SG.Path{..}
-    ensureInvariantsHold n =
-      let allErrors = fmap InvariantViolated (catMaybes [SG.checkEdgeInvariants n, SG.checkNodeInvariants n])
-      in case allErrors of
-        []   -> pure n
-        x:xs -> Left (x :| xs)